+++ src/__internal__/tooltip-provider/index.js
@@ -1,5 +1,6 @@
-import React from "react";
+import React, { useRef } from "react";
 import PropTypes from "prop-types";
+import guid from "../utils/helpers/guid/guid";
 
 export const TooltipContext = React.createContext({});
 
@@ -9,20 +10,27 @@ export const TooltipProvider = ({
   helpAriaLabel,
   focusable = true,
   tooltipVisible,
-  disabled = false,
-}) => (
-  <TooltipContext.Provider
-    value={{
-      tooltipPosition,
-      helpAriaLabel,
-      focusable,
-      tooltipVisible,
-      disabled,
-    }}
-  >
-    {children}
-  </TooltipContext.Provider>
-);
+  disabled,
+  target,
+}) => {
+  const tooltipId = useRef(guid());
+
+  return (
+    <TooltipContext.Provider
+      value={{
+        tooltipPosition,
+        helpAriaLabel,
+        focusable,
+        tooltipVisible,
+        disabled,
+        tooltipId,
+        target,
+      }}
+    >
+      {children}
+    </TooltipContext.Provider>
+  );
+};
 
 TooltipProvider.propTypes = {
   children: PropTypes.node.isRequired,
@@ -31,4 +39,5 @@ TooltipProvider.propTypes = {
   focusable: PropTypes.bool,
   tooltipVisible: PropTypes.bool,
   disabled: PropTypes.bool,
+  target: PropTypes.instanceOf(Element),
 };