+++ src/components/app-wrapper/app-wrapper.component.js
@@ -4,6 +4,9 @@ import classNames from "classnames";
 import tagComponent from "../../__internal__/utils/helpers/tags/tags";
 import { validProps } from "../../__internal__/utils/ether";
 import StyledAppWrapper from "./app-wrapper.style";
+import Logger from "../../__internal__/utils/logger";
+
+let deprecatedWarnTriggered = false;
 
 /**
  * Manages the width and containment of your application.
@@ -12,6 +15,14 @@ class AppWrapper extends React.Component {
   constructor(...args) {
     super(...args);
     this.classes = this.classes.bind(this);
+
+    if (!deprecatedWarnTriggered) {
+      deprecatedWarnTriggered = true;
+      // eslint-disable-next-line max-len
+      Logger.deprecate(
+        "`AppWrapper` component is now deprecated and will soon be removed. Please use the `Box` component instead."
+      );
+    }
   }
 
   /**