+++ src/locales/locale.d.ts
@@ -19,6 +19,16 @@ interface Locale {
   date: {
     dateFnsLocale: () => DateFnsLocale;
   };
+  dialog: {
+    ariaLabels: {
+      close: () => string;
+    };
+  };
+  dialogFullScreen: {
+    ariaLabels: {
+      close: () => string;
+    };
+  };
   errors: {
     messages: {
       formSummary: (
@@ -53,6 +63,14 @@ interface Locale {
     placeholder: () => string;
     noResultsForTerm: (term: string) => string;
   };
+  link: {
+    skipLinkLabel: () => string;
+  };
+  sidebar: {
+    ariaLabels: {
+      close: () => string;
+    };
+  };
   switch: {
     on: () => string;
     off: () => string;
@@ -73,6 +91,11 @@ interface Locale {
   };
   tileSelect: {
     deselect: () => string;
+  };
+  toast: {
+    ariaLabels: {
+      close: () => string;
+    };
   };
   wizards: {
     multiStep: {